Changed USB_Current from float to int16_t, added big_to_native conversions for int16_t

+++ include/tofcore/CommandTypes.hpp
@@ -59,7 +59,7 @@ struct versionData_t
 struct Sensor_Status_t
 {
     int16_t  lastTemperature;
-    float    USB_Current;
+    int16_t  USB_Current;
     uint32_t BIT_Status;
 } PACKED;
